+++ pkg/api/services/messenger/client/types.go
@@ -13,7 +13,7 @@ type IClient interface {
 	GetMessageLimit(context.Context) (uint64, error)
 	ListenChat(context.Context, string, string) (message.IMessage, error)
 	PushMessage(context.Context, string, string) (time.Time, error)
-	LoadMessages(context.Context, string, uint64, uint64, bool) ([]message.IMessage, error)
+	LoadMessage(context.Context, string, uint64) (message.IMessage, error)
 	CountMessages(context.Context, string) (uint64, error)
 }
 
@@ -23,6 +23,6 @@ type IRequester interface {
 	GetMessageLimit(context.Context) (uint64, error)
 	ListenChat(context.Context, string, string) (message.IMessage, error)
 	PushMessage(context.Context, string, string) (time.Time, error)
-	LoadMessages(context.Context, string, uint64, uint64, bool) ([]message.IMessage, error)
+	LoadMessage(context.Context, string, uint64) (message.IMessage, error)
 	CountMessages(context.Context, string) (uint64, error)
 }